# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projeto Picolotec API",
        "description": "Documentação oficial do backend do projeto Picolotec",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/requisitions/": {
            "get": {
                "tags": [
                    "Requisitions"
                ],
                "summary": "Busca todas requisições",
                "operationId": "a830b06c1cf14291bcb9d29d69d05a29",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/RequisitionResource"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/requisitions/{id}": {
            "get": {
                "tags": [
                    "Requisitions"
                ],
                "summary": "Busca uma requisição pelo ID",
                "operationId": "7af6be25288631ffdc85c2b0d8a83762",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/RequisitionResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Requisitions"
                ],
                "summary": "Remove uma requisição",
                "operationId": "f2b113efcc4526d23a7a3ef7055b90e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deletado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResource"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Requisitions"
                ],
                "summary": "Atualiza uma requisição",
                "operationId": "6edfa087d940894958a91ad4a8d0faac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateRequisitionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/RequisitionResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/requisitions": {
            "post": {
                "tags": [
                    "Requisitions"
                ],
                "summary": "Cria uma nova requisição",
                "operationId": "c9dc8010694deae8881387f2eac6dec7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreRequisitionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/RequisitionResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResource"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreRequisitionRequest": {
                "required": [
                    "title",
                    "description",
                    "estimated_value",
                    "urgency",
                    "importance",
                    "user_id"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Compra de Notebook"
                    },
                    "description": {
                        "type": "string",
                        "example": "Necessário para o novo desenvolvedor"
                    },
                    "estimated_value": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 4500.5
                    },
                    "urgency": {
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high"
                        ],
                        "example": "high"
                    },
                    "importance": {
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high"
                        ],
                        "example": "medium"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "UpdateRequisitionRequest": {
                "description": "Esquema para atualização parcial de uma requisição",
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Notebook de Alto Desempenho"
                    },
                    "description": {
                        "type": "string",
                        "example": "Alteração da descrição original"
                    },
                    "estimated_value": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 5200
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "under_review",
                            "approved",
                            "rejected",
                            "completed",
                            "archived"
                        ],
                        "example": "under_review"
                    },
                    "urgency": {
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high"
                        ],
                        "example": "high"
                    },
                    "importance": {
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high"
                        ],
                        "example": "high"
                    }
                },
                "type": "object"
            },
...
            "ErrorResource": {
                "title": "Error Resource",
                "description": "Estrutura padrão de resposta de erro da API",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "code": {
                        "type": "integer",
                        "example": 404
                    },
                    "message": {
                        "type": "string",
                        "example": "Recurso não encontrado."
                    }
                },
                "type": "object"
            },
...
            "RequisitionResource": {
                "title": "Requisition Resource",
                "description": "Estrutura detalhada de retorno de uma requisição",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Compra de suprimentos"
                    },
                    "description": {
                        "type": "string",
                        "example": "Pedido de novos teclados mecânicos"
                    },
                    "estimation_value": {
                        "type": "number",
                        "format": "float",
                        "example": 1500.5
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "priority": {
                        "properties": {
                            "urgency": {
                                "type": "integer",
                                "example": 5
                            },
                            "importance": {
                                "type": "integer",
                                "example": 3
                            },
                            "score": {
                                "type": "integer",
                                "example": 15
                            }
                        },
                        "type": "object"
                    },
                    "permissions": {
                        "properties": {
                            "can_edit": {
                                "type": "boolean",
                                "example": true
                            }
                        },
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "SuccessResource": {
                "title": "Success Resource",
                "description": "Estrutura padrão de resposta de sucesso da API",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "code": {
                        "type": "integer",
                        "example": 200
                    },
                    "message": {
                        "type": "string",
                        "example": "Recurso encontrado."
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}